requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
IPFS-Toolkit==0.5.23
ipfshttpclient==0.7.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.0.5
netaddr==1.2.1
parsimonious==0.10.0
protobuf==5.26.1
pycryptodome==3.20.0
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.1
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
termcolor==2.4.0
toolz==0.12.1
typing_extensions==4.11.0
urllib3==2.2.1
varint==1.0.2
web3==6.19.0
websockets==12.0
yarl==1.9.4

certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.14.0
fsspec==2024.5.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
pillow==10.3.0
requests==2.32.1
scipy==1.13.0
sympy==1.12
torch==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.11.0
urllib3==2.2.1
